company: add selectCompanyFees on reducer
