+++ packages/pilot/src/pages/EmptyState/EmptyState.js
@@ -2,23 +2,15 @@ import React, { useEffect } from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import {
-  always,
-  anyPass,
   applySpec,
   compose,
-  equals,
-  find,
   head,
-  not,
   path,
-  pathOr,
   propOr,
-  pluck,
   pipe,
   prop,
   propEq,
   split,
-  when,
 } from 'ramda'
 import { translate } from 'react-i18next'
 import EmptyStateContainer from '../../containers/EmptyState'
@@ -29,6 +21,8 @@ import {
 } from './actions'
 import isOnboardingComplete from '../../validation/isOnboardingComplete'
 
+import { selectCompanyFees } from '../Account/actions/reducer'
+
 const getUserName = pipe(prop('name'), split(' '), head)
 
 const hasAdminPermission = propEq('permission', 'admin')
@@ -38,38 +32,7 @@ const getAccessKeys = applySpec({
   encryptionKey: path(['encryption_key', environment]),
 })
 
-const getAntifraudCost = pipe(
-  pathOr([], ['gateway', environment, 'antifraud_cost']),
-  find(propEq('name', 'pagarme')),
-  prop('cost')
-)
-
 const getAlreadyTransacted = propOr(true, 'alreadyTransacted')
-
-const notDefaultInstallments = pipe(
-  pluck('installment'),
-  anyPass([equals([1, 2, 7]), equals([1])]),
-  not
-)
-
-const getInstallmentsFee = pipe(
-  pathOr([], ['psp', environment, 'mdrs']),
-  find(propEq('payment_method', 'credit_card')),
-  pathOr([], ['installments']),
-  when(notDefaultInstallments, always([]))
-)
-
-const getFees = pipe(
-  prop('pricing'),
-  applySpec({
-    anticipation: path(['psp', environment, 'anticipation']),
-    antifraud: getAntifraudCost,
-    boleto: path(['gateway', environment, 'boletos', 'payment_fixed_fee']),
-    gateway: path(['gateway', environment, 'transaction_cost', 'credit_card']),
-    installments: getInstallmentsFee,
-    transfer: path(['transfers', 'ted']),
-  })
-)
 
 const mapDispatchToProps = {
   requestOnboardingAnswers: requestOnboardingAnswersAction,
@@ -86,7 +49,7 @@ const mapStateToProps = ({
 }) => ({
   accessKeys: getAccessKeys(company),
   alreadyTransacted: getAlreadyTransacted(company),
-  fees: getFees(company),
+  fees: selectCompanyFees(company),
   isAdmin: hasAdminPermission(user),
   isMDRzao: company && propEq('anticipationType', 'MDRZAO', company),
   onboardingAnswers,
